fix: 댓글 목록 컴포넌트에서는 데이터 받아서 처리하도록 수정
